chore(Camel): Extract isCamelRoute and isCamelFrom utils
