+++ packages/ui/src/utils/is-camel-from.ts
@@ -1,0 +1,15 @@
+import { FromDefinition } from '@kaoto/camel-catalog/types';
+import { isDefined } from '@kaoto/forms';
+
+/** Very basic check to determine whether this object is a Camel From */
+export const isCamelFrom = (rawEntity: unknown): rawEntity is { from: FromDefinition } => {
+  if (!isDefined(rawEntity) || Array.isArray(rawEntity) || typeof rawEntity !== 'object') {
+    return false;
+  }
+
+  const objectKeys = Object.keys(rawEntity);
+  const isFromHolder = objectKeys.length === 1 && objectKeys[0] === 'from';
+  const isValidUriField = typeof (rawEntity as { from: FromDefinition })?.from?.uri === 'string';
+
+  return isFromHolder && isValidUriField;
+};
